requirements.txt
Flask==2.2.5
flask-cors==4.0.1
tensorflow==2.18.0
pillow==9.5.0
numpy==1.26.4
gunicorn==20.1.0